adding the global blog posts
