+++ .prettierrc
@@ -1,4 +1,5 @@
 {
     "printWidth": 80,
-    "singleQuote": true
+    "singleQuote": true,
+    "trailingComma": "all"
 }
