+++ src/components/Blog/index.js
@@ -10,38 +10,52 @@ function Blog() {
   const [offset, setOffset] = useState(0);
   const [count, setCount] = useState(0);
   const [tag, setTag] = useState(null);
-
   useEffect(() => {
-    fetch(`http://localhost:4000/api/v1/blog?limit=5&offset=${offset}`, {
-      method: 'GET',
-      mode: 'cors',
-      cache: 'no-cache',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-    })
-      .then((response) => response.json())
-      .then((result) => {
-        setBlogPosts(result.data);
-        setCount(result.itemsCount);
-      })
-      .catch((err) => console.error(err.message));
+    const fetchData = async () => {
+      const response = await fetch(
+        `http://localhost:4000/api/v1/blog?limit=5&offset=${offset}`,
+        {
+          method: 'GET',
+          mode: 'cors',
+          cache: 'no-cache',
+          headers: {
+            'Content-Type': 'application/json',
+          },
+        },
+      );
+      const result = await response.json();
+      setBlogPosts(result.data);
+      setCount(result.itemsCount);
+    };
+    fetchData();
   }, [offset]);
   return (
     <>
-      <Wrapper>
-        <Feed tag={tag} setTag={setTag} />
-        {blogPosts.map((post) => (
-          <Post data={post} key={post.id} />
-        ))}
-        <Pages setOffset={setOffset} offset={offset} count={count} />
-      </Wrapper>
-      <Tags
-        setBlogPosts={setBlogPosts}
-        setCount={setCount}
-        offset={offset}
-        setTag={setTag}
-      />
+      {blogPosts ? (
+        <>
+          <Wrapper>
+            <Feed
+              tag={tag}
+              setTag={setTag}
+              setBlogPosts={setBlogPosts}
+              setCount={setCount}
+              offset={offset}
+            />
+            {blogPosts.map((post) => (
+              <Post data={post} key={post.id} />
+            ))}
+            <Pages setOffset={setOffset} offset={offset} count={count} />
+          </Wrapper>
+          <Tags
+            setBlogPosts={setBlogPosts}
+            setCount={setCount}
+            offset={offset}
+            setTag={setTag}
+          />
+        </>
+      ) : (
+        <p>Loading</p>
+      )}
     </>
   );
 }
